Рефакторинг: генерация карточек
