+++ src/index.pug
@@ -11,7 +11,7 @@ html
         include ./blocks/player/player.pug
         include ./blocks/measurement/measurement.pug
     script(src="https://code.jquery.com/pep/0.4.3/pep.js")
-    script(src="scripts/card-generator.js" type="module")
+    script(src="scripts/card-main.js" type="module")
     script(src="scripts/gestures-camera.js" type="application/javascript")
     script(src="scripts/nav.js" type="application/javascript")
     script(src="scripts/sidebar.js" type="application/javascript")
